新增註冊程式, db修改users index

+++ WordCard/app/src/main/java/edu/harvard/cs50/wordcard/dao/LessonsDao.java
@@ -19,6 +19,6 @@ public interface LessonsDao {
     long insertLesson(Lessons lessons);
 
     @Update
-    long updateLesson(Lessons lessons);
+    int updateLesson(Lessons lessons);
 
 }
