+++ WordCard/app/src/main/java/edu/harvard/cs50/wordcard/dao/WordsDao.java
@@ -19,5 +19,5 @@ public interface WordsDao {
     long insertWord(Words words);
 
     @Update
-    long updateWord(Words words);
+    int updateWord(Words words);
 }
